use block and delete tests for formatter
