+++ spec/rack/camel_snake_spec.rb
@@ -67,29 +67,4 @@ describe Rack::CamelSnake do
     end
   end
 
-  describe 'formatter' do
-    using Rack::CamelSnake::Refinements
-
-    let!(:snake_hash){ { 'is_done' => 'hoge', 'order' => 1, 'task_title' => 'title' } }
-    let!(:camel_hash){ { 'isDone'  => 'hoge', 'order' => 1, 'taskTitle'  => 'title' } }
-    let(:snake_array){ [ snake_hash, snake_hash, snake_hash ] }
-    let(:camel_array){ [ camel_hash, camel_hash, camel_hash ] }
-    let(:to_camel){ ->(key){ key.is_a?(String) ? key.to_camel : key } }
-    let(:to_snake){ ->(key){ key.is_a?(String) ? key.to_snake : key } }
-
-    context 'given :to_camel' do
-      it 'converts keys into camelCase, and the keys should be a string.' do
-        Rack::CamelSnake.formatter(snake_hash,  to_camel).should eq camel_hash
-        Rack::CamelSnake.formatter(snake_array, to_camel).should eq camel_array
-      end
-    end
-
-    context 'given :to_snake' do
-      it 'converts keys into snake_case, and the keys should be a symbol.' do
-        Rack::CamelSnake.formatter(camel_hash,  to_snake).should eq snake_hash
-        Rack::CamelSnake.formatter(camel_array, to_snake).should eq snake_array
-      end
-    end
-  end
-
 end
